On CHERI revoke free'd pointers before reuse
With Cornucopia revocation enabled, free'd pointers reside in
quarantine until revocation occurs.  Use of such pointers is considred
mostly harmless, but is not detected as expected by the tests.
Inserting a call to malloc_revoke() ensures that all copies of the
pointer are rendered invalid and that reuse results in a fault.